refactor(ui): extract Overview tab to ui/tabs/overview.rs
Lifts draw_overview, draw_connections_list (~340 lines),
draw_grouped_connections_list (~350 lines), draw_stats_panel
(~420 lines), draw_interface_stats_with_graph (~170 lines), the
DrawContext bundle, and the render_section_separator helper into
ui/tabs/overview.rs (~1,360 lines). Cross-tab cell-style helpers
(state_color, dpi_color, bandwidth_line, status_indicator_cell)
and NONE_PLACEHOLDER bump to pub(crate) so overview and the
upcoming details extraction can reach them.
